docker-compose/Database_Docker/requirements.txt
Flask==1.1.2
psycopg2-binary==2.8.6
Pandas
numpy
scikit-learn
seaborn
imbalanced-learn
joblib
